コンポーネントファイル作成
