+++ frontend/src/components/QuizPushPull.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+const QuizPushPull = ({ quiz, onAnswer, ...props }) => {
+  // このカテゴリ専用のロジックやUIをここに書く
+  return (
+    <div className="...">
+      <h2 className="text-xl font-bold mb-2">押し引き問題</h2>
+      {/* クイズ内容や回答ボタンなど */}
+    </div>
+  );
+};
+
+export default QuizPushPull;
